correct polygon area calculation
When calculating a part of the polygon area and thus ending the polygon
at the specified vertex aEnd we  need to calculate and add the area below
the needed additional line from "aEnd" to "this" to close the resulting
partial polygon. As in this case VERTEX* p references "aEnd" after exiting
the do-while loop the correction  term in the if-statement was always
evaluated to zero. Change "aEnd" to "this" in the correction term.