Maintain std::move to avoid copy penalty
Caching lib models is a slow point when loading libraries when many
footprints have embedded models.  We want to speed this by not copying
data unneccesarily.  Some compilers will invalidate the models structure
right-to-left in the function call (the order is unspecified by the
standard), leading to broken model names unless we cache the name before
hand.