Remove Context Menu in Empty Space of Hierarchy Navigator
Previously, a single RMB-click at empty space had no effect, while a
double-click opened the context menu. As per the @sethhillbrand suggestion,
instead of fixing this, no context menus should appear in empty space, as
it is not necessary. To address this, the wxEVT_RIGHT_UP event binding has
been removed by unbinding it from both the constructor and destructor.
Additionally, the onRightClick(wxMouseEvent& aEvent) method has been
removed, as it is linked to wxEVT_RIGHT_UP and is no longer necessary.

Fixes #17962