Pcbnew: initialise ZONE_SETTINGS::m_ruleAreaPlacementSourceType
Fixes a text failure in qa_pcbnew.

If this is not initialised, at least the comparison operator will be
operating on junk member values if either side has not had the
source type set.

Maybe it doesn't even make sense for the source type/source to be
checked if m_isRuleArea is false (they're presumably "do not care" in
that context), but if they're initialised, it doesn't matter either
way.