Pcbnew: Avoid creating a footprint in a read-only library
When using "New Footprint" from a selected library in the tree view
(or with an open footprint in such a library) that is read-only,
the new footprint will now be created in a "non-library-assigned" state
so the user has to Save As to save it to a library.

There will also be a warning when this happens to (try to) be less
mysterious about what just happened.

This patch uses a similar logic to that provided by Jan Wichmann in
MR !2034 - thank you!

Relates-To: https://gitlab.com/kicad/code/kicad/-/issues/18923