Pcbnew dimension: remove misleading m_isDiamter in RADIAL
Presumably the intention was that the RADIAL dimension could
also be a diameter type of dimension. However, as the recent
inward/outward work showed, this is actually rather fraught,
as a diameter type needs a different prefix (so now you can't
always tell when it's user-set or not) and there are different
numbers of control points.

Possibly, a real "Diameter" type should be a new class - either
a subclass of this one or another one entirely.

In the meantime, remove the m_isDiameter member, as it's never
set to anything other than 'false' and inaccessible.